# contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: TUBA Product API
  version: 0.1.0
  description: Current API contract. Operations marked planned are design contracts and are not implemented endpoints.
servers:
  - url: /api/v1
security:
  - oidc: []
paths:
...
  /anomalies/{id}:
    get:
      operationId: getAnomaly
      x-required-permission: anomaly:read
      summary: Get a tenant-scoped anomaly
      parameters: [{name: id, in: path, required: true, schema: {type: string, maxLength: 256}}]
      responses:
        "200":
          description: Anomaly detail
          content: {application/json: {schema: {$ref: "#/components/schemas/AnomalyDetail"}}}
        "404": {description: Anomaly not found}
...
components:
  securitySchemes:
    sourceApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    collectorBearer:
      type: http
      scheme: bearer
      bearerFormat: TUBA Collector credential
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example/realms/tuba/.well-known/openid-configuration
...
  schemas:
...
    AnomalySummary:
      type: object
      required: [id, type, severity, status, timestamp, entity, evidence_count]
      properties:
        id: {type: string}
        type: {type: string}
        severity: {type: string, enum: [low, medium, high, critical]}
        status: {type: string, enum: [open, investigating, closed, false_positive]}
        timestamp: {type: string, format: date-time}
        score: {type: number}
        entity:
          type: object
          required: [id, type]
          properties:
            id: {type: string}
            type: {type: string}
        rule_id: {type: string}
        rule_version: {type: string}
        summary: {type: string}
        evidence_count: {type: integer}
    AnomalyDetail:
      allOf:
        - $ref: "#/components/schemas/AnomalySummary"
        - type: object
          required: [reason_codes, evidence_event_ids]
          properties:
            reason_codes: {type: array, items: {type: string}}
            evidence_event_ids: {type: array, items: {type: string}}
...